recommender/requirements.txt
aenum==2.1.2
altgraph==0.10.2
aniso8601==6.0.0
bdist-mpkg==0.5.0
bonjour-py==0.3
certifi==2019.3.9
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.7
fuzzywuzzy==0.17.0
gunicorn==19.9.0
hunlp==0.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
jprops==2.0.2
macholib==1.5.1
MarkupSafe==1.1.1
matplotlib==1.3.1
modulegraph==0.10.4
mysql-connector-python==8.0.16
nltk==3.4
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
py2app==0.7.3
pyobjc-core==2.5.1
pyobjc-framework-Accounts==2.5.1
pyobjc-framework-AddressBook==2.5.1
pyobjc-framework-AppleScriptKit==2.5.1
pyobjc-framework-AppleScriptObjC==2.5.1
pyobjc-framework-Automator==2.5.1
pyobjc-framework-CFNetwork==2.5.1
pyobjc-framework-Cocoa==2.5.1
pyobjc-framework-Collaboration==2.5.1
pyobjc-framework-CoreData==2.5.1
pyobjc-framework-CoreLocation==2.5.1
pyobjc-framework-CoreText==2.5.1
pyobjc-framework-DictionaryServices==2.5.1
pyobjc-framework-EventKit==2.5.1
pyobjc-framework-ExceptionHandling==2.5.1
pyobjc-framework-FSEvents==2.5.1
pyobjc-framework-InputMethodKit==2.5.1
pyobjc-framework-InstallerPlugins==2.5.1
pyobjc-framework-InstantMessage==2.5.1
pyobjc-framework-LatentSemanticMapping==2.5.1
pyobjc-framework-LaunchServices==2.5.1
pyobjc-framework-Message==2.5.1
pyobjc-framework-OpenDirectory==2.5.1
pyobjc-framework-PreferencePanes==2.5.1
pyobjc-framework-PubSub==2.5.1
pyobjc-framework-QTKit==2.5.1
pyobjc-framework-Quartz==2.5.1
pyobjc-framework-ScreenSaver==2.5.1
pyobjc-framework-ScriptingBridge==2.5.1
pyobjc-framework-SearchKit==2.5.1
pyobjc-framework-ServiceManagement==2.5.1
pyobjc-framework-Social==2.5.1
pyobjc-framework-SyncServices==2.5.1
pyobjc-framework-SystemConfiguration==2.5.1
pyobjc-framework-WebKit==2.5.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
PyPika==0.28.0
python-dateutil==2.8.0
pytz==2019.1
rake-nltk==1.0.4
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.2
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
vboxapi==1.0
Werkzeug==0.15.0
xattr==0.6.4
